+++ src/dlong/__init__.py
@@ -1,4 +1,5 @@
-# flake8: noqa
-from .core import QuadraticDamageModel
-from .core import FractionalDiscount
-from .core import ExampleRecipe
+# noqa: F401
+from dlong.core import QuadraticDamageModel
+from dlong.core import FractionalDiscount
+from dlong.core import ExampleRecipe
+from dlong._version import __version__